pikkujoulu_ui/requirements.txt
Flask-SocketIO
Flask-MQTT
Flask-Bootstrap4
Flask
paho-mqtt
requests
gevent
gunicorn
pytest
pytest-runner
